trait moved to controller
